+++ CHANGELOG.md
@@ -57,6 +57,7 @@ All notable changes to this project will be documented in this file.
 - Fixed MultiPHP upgrade script to update all web templates.
 - Fixed report issue link in installer scripts.
 - Fixed database user authentification on backup restore.
+- Added robots.txt for roundcube webmail to prevent search bot crawling.
 
 ## [1.0.6] - 2019-09-24 - Hotfix
 ### Bugfixes